refactor: drop quic and use net.Pipe for testing
