+++ go.mod
@@ -10,10 +10,9 @@ require (
 	github.com/mattn/go-sqlite3 v1.14.7
 	github.com/peterh/liner v1.2.1
 	github.com/pkg/errors v0.9.1
-	github.com/quic-go/quic-go v0.34.0
 	github.com/spf13/cobra v1.2.1
 	github.com/stretchr/testify v1.7.0
-	golang.org/x/sync v0.1.0
-	golang.org/x/sys v0.5.0
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
+	golang.org/x/sys v0.0.0-20210616094352-59db8d763f22
 	gopkg.in/yaml.v2 v2.4.0
 )
